+++ apps/fe/src/components/common/footer/index.ts
@@ -1,0 +1,1 @@
+export { default as OverlayFooter } from "./OverlayFooter";
