+++ apps/fe/src/components/common/index.ts
@@ -1,5 +1,5 @@
-export { default as Button } from "./Button/Button";
-export type { ButtonProps } from "./Button/Button";
-export { default as OverlayFooter } from "./footer/OverlayFooter";
-export { default as SectionHeader } from "./section-header/SectionHeader";
-export { default as Skeleton } from "./skeleton/Skeleton";
+export { Button } from "./Button";
+export type { ButtonProps } from "./Button";
+export { OverlayFooter } from "./footer";
+export { SectionHeader } from "./section-header";
+export { Skeleton } from "./skeleton";
