+++ apps/fe/src/components/common/section-header/index.ts
@@ -1,0 +1,1 @@
+export { default as SectionHeader } from "./SectionHeader";
