+++ apps/fe/src/components/common/skeleton/index.ts
@@ -1,0 +1,1 @@
+export { default as Skeleton } from "./Skeleton";
